"Calculadora de Propinas"

+++ Class5/src/Converter.java
@@ -6,7 +6,7 @@ public class Converter {
         Scanner milesTotal = new Scanner(System.in);
 
         System.out.print("\nIngrese las millas a convertir: ");
-        double miles = milesTotal.nextDouble();
+        float miles = milesTotal.nextFloat();
 
         double kilometres = (miles * 1.60934);
         System.out.println("La distancia en Kilómetros es: " + kilometres);
